+++ RELEASE NOTES.md
@@ -8,11 +8,20 @@
 1. Auto deploy when group config screen is exited
 1. Calendar view for multiple playlists scheduling under Group
 1. Link edit
+1. Availability of beta version
 
-- 1.6.0
+**1.6.0**
 1. customized logo and url (directly change at public/app/img/pisignage-logo.png and public/app/partials/menu.jade)
 2. UI changes (specific to pisignage.com)
 3. editing of links
 4. Showing playlists associated with Asset in Asset List and Edit screens, thumbnail in Edit assets
 5. License Status, player name change, location display
 
+**1.7.0**
+1. Assign asset directly to multiple playlists in asset details screen and assets list screen
+2. Deploy All button
+3. warn user to deploy in playlist
+4. deploy to groups from playlist
+5. Player MAC address display on the server page
+6. font download from Google for offline use
+7. tooltips